Fixed and cleaned up logic for selecting the last used theme.
For some reason, selecting the last used theme recently stopped working 
for some games if the theme was the default one. I'm not sure, but 
this might be related to aseigo's changes to KConfigSkeleton and default 
values.

Regardless the new code should work the same as before, with the added 
bonus that the "Theme" KConfig entry is no longer hardcoded to the
"General" group. As long as it's named "Theme", it will be found
automatically.

There is still a potential issue when running a game for the first time.
As a game developer there are a couple things you can do to prevent 
this. One is to set the default value of the "Theme" entry to "themes/
default.desktop" (or whatever your default theme is named). The other is 
to update the KConfigSkeleton whenever you fallback to default theme.
For example:
	if ( !Render::loadTheme( Settings::theme() ) )
	{
		Render::loadDefaultTheme();
		Settings::setTheme("themes/default.desktop");
	}

CCMAIL:kde-games-devel@kde.org


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@868325 283d02a7-25f6-0310-bc7c-ecb5cbfe19da