Major rewrite to the animation interaction between the Engine and Scene. There were a couple of minor bugs with the animation sequencing that I found I couldn't fix without breaking something else in the old system.
The Engine now queues up a number of animation "stages" in the Scene and 
only triggers the animation to start when necessary. This involves a bit 
more code in the Scene and a bit less in the Engine, but the result is a 
whole lot less fragile and a whole lot more readable.

This change has most likely introduced some regressions. More cleanup 
and changes to follow.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@886308 283d02a7-25f6-0310-bc7c-ecb5cbfe19da