rework pause mode
- game is now paused if the board loses focus (e.g. if a menu is opened
        or the application itself loses focus)
- game can now be resumed by clicking the board

There is a strange timer issue that needs to be resolved.
If you pause the game and then choose to play a new game, the timer has
the negative pause time. o.O
But that's a different matter and the timers are on my list anyway. Too
many type casts... :)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@923457 283d02a7-25f6-0310-bc7c-ecb5cbfe19da