Fixed functionality that allowed one to jump directly to the section for the current game type in KHelpCenter. Previously the action was assigned the shortcut F1 which caused a conflict with the default "KPatience Handbook" action.
This commit changes the shortcut to F2 and also adds the action to the Help menu. The action texts is "Help with <game>" where <game> is dynamically set to the name of the current game.

This commit also simplifies the code use to determine the docbook anchor of the current game and fixes and issue with the ampersand in the "Forty & Eight".

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@945172 283d02a7-25f6-0310-bc7c-ecb5cbfe19da