unpause the game if the player requests a hint
There is a problem with the current pause-on-focus-loss behaviour.
If the player opens the menu to request a hint, the hint action is
disabled due to the pause mode.
I now keep the hint action enabled in pause mode and unpause the game
automatically if the player requests a hint.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@965416 283d02a7-25f6-0310-bc7c-ecb5cbfe19da