Fixed a Q_ASSERT based crashed in the Grandfather's Clock solver.
The translate_layout() method asserted that each of the 12 target piles always contain at least one pile. Unfortunately, if it takes an especially long time to render the card graphics at the start of a new game the 250ms solver restart timer can time out and this method can be called before the intitial deal has even happened.

The proper fix would be to ensure that the solver is never started before the cards have actually been dealt, but I saw no way of cleanly implementing that.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@967673 283d02a7-25f6-0310-bc7c-ecb5cbfe19da