Some new card decks have entered into SVN, which although very beautiful there also quite large. For example, the Ancient Egyptians card deck adds 3 seconds to KPat's start up time compared to a lighter deck like Nicu White. This increase in start up time isn't due to rendering time, as KPat already makes use of KCardCache to cache the rendered pixmaps between runs. The time consumed is used purely to load the SVG file into the KSvgRenderer.
This patch to KCardCache causes the SVG file to only be loaded on into the KSvgRenderer on first use. At first I thought this would be sufficient to improve start up time when the cards are already cached, but it is not. KPat must also call defaultFrontSize() on start up to learn the aspect ratio of the cards (something LSkat currently doesn't do). Previously, this function used to get the element size from the renderer, which then has to be loaded. This patch also causes that function to store it's default size in the cache, eliminating the need to access the renderer the next time the function is run, even across sessions.

I should point out that the expense of loading the SVG hasn't been removed, just delayed until a card rendering is needed that is already cached. This ended up being an improvement in some situations, but an annoyance in others. For example, having the load time occur on start up is much nicer than having it occur the first time you flip a card in Klondike.

As a result, I looked using KCardCache's built in threaded loading. Unfortunately, that logic had some issues. The biggest of these was that it tried to use QPixmaps in a non GUI thread. The "dokill" boolean was also uninitialised on creation, so threads often decided to kill themselves for no reason. (That one took a while to figure out.)

This updated patch contains a working threaded loading implementation. Support for threaded loading of the backside has been dropped as it would have required a lot of additional code zero performance gain. (The backside is almost always loaded in the GUI thread and the time to load the small backside SVG is neglible.)

I also have a six line patch to KPat that adds threaded loading. The result is a much faster response on startup, resize and deck change, even on my single core processor. I would imagine it would be even slicker on a multicore system. A similar patch for LSkat should be easy enough.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@968838 283d02a7-25f6-0310-bc7c-ecb5cbfe19da