Reworked and cleaned up a fair bit of the scene layout code. More work to come.
Managed to finally fix the inverted pile bug when playing Spider. Unfortunately the change is rather large and somewhat invasive, so I don't think it makes sense to backport to the 4.3 branch. The risk of breakage is just to high for such an easily worked around, tricky to reproduce display issue.

BUG:201683

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1009850 283d02a7-25f6-0310-bc7c-ecb5cbfe19da