Added the concept of a contentsArea to the layout handling.
All piles are now arranged inside of a contents area, which gives us better options for positioning the entire playarea withing the view. The contents area always starts at 0,0 which simplifies positioning code. The area is centred on the screen by applying a negative offset to the sceneRect. Items with negative x positions are no longer thrown way off to the right of the view.

The margin around the cards is now handle in the sceneRect, so it needn't be implemented separately in each DealerScene.

The concept of a the contentsArea will also allow us to easily add new vertical and horizontal sizing and positioning policies.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1010651 283d02a7-25f6-0310-bc7c-ecb5cbfe19da