Improvements to the Mod3 layout.
Set layout spacing to 0 to prevent the very close together piles from interfering with one another.

Set a very low spread for the first three rows. This lets you quickly distiguish the numer of cards in the pile by height, which is handy. Also we don't care about the cards below, so there's no real reason to show much of them.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1011069 283d02a7-25f6-0310-bc7c-ecb5cbfe19da