Further clean up of class interdependencies.
DealerScene subclasses no longer find the MainWindow instance and insert their actions into the menu themselves. Instead they have a getter that returns the list, which the mainwindow then inserts itself.

As a result, View no longer needs to keep track of its instance or the instance of the MainWindow that contains it. This means even more code removed from View.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1012201 283d02a7-25f6-0310-bc7c-ecb5cbfe19da