Fixed a bug where the title bar didn't relfect the new game number when changing Klondike or Spider options.
Previously, when changing the options in Klondike or Spider, a new random game was opened, but the game number shown in the titlebar wasn't updated. Getting this to work correctly would be tricky, so I took the lazy route and made it just reload the same game number. No need to update the titlebar!

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1030625 283d02a7-25f6-0310-bc7c-ecb5cbfe19da