Removed AUTHORS, CHANGES, README, README.tkcTrump and TODO.
These files are like hidden relics from the early days of KDE. README in particular is a perfectly preserved artifact, completely untouched for the last eleven and a have years.

The AUTHORS file has since been replaced with the credits in the About dialog. TODO and CHANGES are so out of date it isn't funny. Apparently tkcTrump was a version of KPat ported to PDAs.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1030964 283d02a7-25f6-0310-bc7c-ecb5cbfe19da