CardDeck no longer inherits from Pile.
CardDeck being both a partial singleton and a QGraphicsItem made things tricky as it would survive beyond the life of its QGraphicsScene. This also meant we had to useless pile kicking around in games that didn't use a talon.

CardDeck now has two main tasks. It creates, shuffles and dispenses Cards. It serves as the central location for Card geometries and pixmaps.

This is a very major change, but so far seems to be working fine. More polish and testing will be needed before I can move it in to trunk, though.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1031082 283d02a7-25f6-0310-bc7c-ecb5cbfe19da