Changed the Card transformationMode to SmoothTransformation.
This corrects and issue that has been showing up in recent KPat trunk. When right clicking on a card to make it "pop up" for inspection the card graphics became very grainy with some obvious scaling artifacts. I haven't yet deduce what caused this change, but this fixes it. SmoothTransform is more expensive, but the Cards are transformed so rarely, I don't think this is an issue.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1032480 283d02a7-25f6-0310-bc7c-ecb5cbfe19da