Initial port to QtKinetic.
First stab at using the new animation framework in Qt4.6. This is mostly a straight port, with the QGraphicsItemAnimation code replaced with its rough QAnimation equivalent. There's still plenty of room for improvements.

The performance gains are considerable. I always new that each card having it's own QTimeline was causing unsynced painting, but I had no idea the effect was so significant. The deal animations are now super smooth. :D

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1039749 283d02a7-25f6-0310-bc7c-ecb5cbfe19da