Removed the "getUp" functionality.
Previously, if you right clicked on a face down card, for some (intentional?) reason it would jump to the top for 1.5 seconds. I have no idea why it did this, but it always seemed broken to me.

Now right clicking on a facedown card will use the same zoom/slide/rotate animation as face up cards. Is this useful? Not really, but who am I to tell which cards you can zoom/slide/rotate and which you can't.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1040150 283d02a7-25f6-0310-bc7c-ecb5cbfe19da