New feature in KAtomic: support for level packs.
Feature was developed in a separate branch, this is a merge to trunk.

From now on KAtomic will support level packs - collection of levels.
Default levels were converted to reside in single level pack file.

Update procedure should work seamlessly:
- I've added a kconf_update script to ensure that all current user settings
will be converted to a new format (which changed slightly now to include info about levelpack)
- Old save files should work too.
- Highscores are now in a separate file in a data dir, old ones should be converted

Conversion will happen on a first start of KAtomic after this commit.

Ah, one thing that is still in TODO state is a support for GHNS.
I need to catch up with ghns guys for setting up server side, then adding client side support
should be really easy.

And as the feature is otherwise completed I want to merge it right now, so it has a chance to be tested :)


Enjoy and DO test this please :)
C&C are welcomed too.

(i think i'll close bug after ghns will be there and working)

CCBUG: 205557



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1040387 283d02a7-25f6-0310-bc7c-ecb5cbfe19da