Added an static cache of KCardCaches to CardDeck.
Constructing new KCardCaches is expensive, because it requires reloading the SVG into the QSvgRenderer. CardDeck now keeps up the three last used KCardCaches on hand to avoid having to construct new ones.

Doing this in a static private class means CardDeck doesn't have to be a singleton and instances can be created and destroyed when needed without worrying about losing the cache.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1041439 283d02a7-25f6-0310-bc7c-ecb5cbfe19da