Remove the spread() and setSpread() methods from Card.
Caching the layout spreads in the cards themselves seems like a good idea
at first, but in reality there is very little gain for the cost of having
to always make sure the spread is up to date. Spreads are now calculated
when needed using cardOffset(). If some future Pile subclass needs more
advanced functionality, cardOffset() is virtual, so it can be easily
reimplemented.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1044898 283d02a7-25f6-0310-bc7c-ecb5cbfe19da