Improve run complete animation in Spider.
In Spider, the player builds runs of cards from king (on the bottom) to
ace (on the top). Once a run is complete, the cards are moved to a new
pile on the lower left part of the screen. Previously, this new pile
ended up with the cards reversed (ace on bottom, king on top). That's
weird. A real player, with real cards would never do this. As a result
the animation required the cards to pass through one another in order to
reverse the order and that looked bad.

Now the ace remains on top. If there is some great historical reason why
the king was on top, I guess I could change it back.

Also, the speed of this animation is now controled by speeds.h.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1044927 283d02a7-25f6-0310-bc7c-ecb5cbfe19da