Improve the tracking of whether animations are in progress.
Now that all animations pass through a single function, we can easily
keep track of animations and prevent input while that's happening.

The tracking has moved to CardDeck. I'm not sure this is the most
logical place for this, but it was by far the easiest to implement. More
changes might happen.

This fixes several regressions, but probably introduces a few as well.
It looks good so far, though.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1046560 283d02a7-25f6-0310-bc7c-ecb5cbfe19da