Created a generic CardScene class to sit between QGraphicsScene and DealerScene.
DealerScene is a huge class. It currently handles a lot of tasks some
of which aren't necessarily that related. The goal of the CardScene
class is twofold. First, to to move the generic code related a scene of
Cards and Piles out of DealerScene, allowing it to focus on more
KPat specific tasks. Second, to create a generic class that might be
usable for future KDE card games

It doesn't do anything currently, other than compile. The license and
copyright don't make sense currently, but will once existing code has
been moved from DealerScene to CardScene.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1052555 283d02a7-25f6-0310-bc7c-ecb5cbfe19da