Work around bug 215136.
The demo mode was getting stuck, because calling translateMove on the
first element of minMoves returned 0. In such a case, I think it makes
sense to fallback on one of the other hints, if present.

I'm not closing the bug, as this merely patches over the root problem.
Obviously if I move is the list of winning moves, it should be valid
and translatable.

CCBUG: 215136

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1057664 283d02a7-25f6-0310-bc7c-ecb5cbfe19da