Do not process mouse events while the connection is drawn.
Clicking on one of the already connected tiles would have selected
it before removing it, leading to a situation where the player could
not select another pair of tiles.

This is more a workaround than a proper fix but I have to understand
the usage of m_paintConnection first in order to consider its reusage
here.

The downside is that you cannot continue the game while the connection
is shown. I consider this the lesser of the two evils since you can
set the timeout

Thanks Marcel for the report.

BUG: 221498

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1072296 283d02a7-25f6-0310-bc7c-ecb5cbfe19da