Create KCardThemeWidget, a fancy new card deck selection widget.
It uses Qt's ModelView framework to make a rather slick looking item
list. Previews are rendered directly from the SVG, so there's no need
to keep PNG previews around. Each app can specify which cards it wishes
to see in the preview.

This required copying CardDeckInfo into LibKCardGame. This is only
temporary until that code is replaced as well. :)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1086700 283d02a7-25f6-0310-bc7c-ecb5cbfe19da