Improve initial widget load time by spreading out rendering.
With a completely empty cache the intitial load time of KCardThemeWidget
was attrocious even on a speedy PC, due to the expense of rendering the
previews. Ideally, we would just render the previews in a worker thread,
but that isn't possible here, as the render logic requires QPixmaps, and
we can't get around that by using QImages in this case.

So the compromise solution is to render a preview in the main thread,
wait 400ms using a single shot QTimer and repeat. This isn't great, but
the dialog appears quickly and the UI remains semi-responsive during the
loading, which takes several seconds even on a fast PC. The consolation,
of course, is that everything is cached after the first run, so the user
will hopefully only see this once.

Not a perfect solution, but I think it's tolerable.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1086701 283d02a7-25f6-0310-bc7c-ecb5cbfe19da