Port configuration to KConfigXT.
I've been meaning to do this for quite a while. KConfigXT is keeps the
config details out of the way and makes KPat more consistent with other
KDE apps. This change should be rc file compatible (despite the
temptation to rename some keys and to move some keys to different
groups.) No user settings should be lost upon upgrading.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1086726 283d02a7-25f6-0310-bc7c-ecb5cbfe19da