Get rid of KCardPile::relayoutCards().
This was an unusual function that delayed the layout until all cards in
the pile were done animating. Gradually over time it became used in
fewer and fewer places. As far as I can tell, none of the existing uses
required that functionality, so it's now gone.

Semi-related, the flipping of the top card has moved from moveCards() to
layoutCards().

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1086728 283d02a7-25f6-0310-bc7c-ecb5cbfe19da