Add new private KCardAnimation to replace complex QPropertyAnimations.
Rather than set up a bunch of properties on KCardPrivate with a bunch of
pass through methods, just to connect to a QParallelAnimationGroup of
QPropertyAnimations, it's much cleaner, simpler and less code to
implement a custom animation class. Because KCardAnimation is so
specialised, it has very little code and may even be more performant.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1086741 283d02a7-25f6-0310-bc7c-ecb5cbfe19da