Fix some warnings about Cards being removed multiple times.
Cards were being removed from the scene after win animation to prevent
them from showing up on the edges if the window was resized. Then if the
CardDeck was destroyed before a new game was started, there was an
attempt to remove each of the cards again, resulting in a slew of
warnings from Qt.

Now we just hide the cards instead of removing them. This makes more
sense, anyway.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1105318 283d02a7-25f6-0310-bc7c-ecb5cbfe19da