Split KCardPile::add(KCard,index) into separate insert and add methods.
The previous add function was always an ugly part of the pile API. If
the default index of -1 was used (as it was in the majority of cases) it
would simply add the card to the top of the pile. If an explicit index
was given and that index was less than the size of the pile, the card at
the index would be overwritten. If the index was greater than the size,
null card pointers would be appended to the pile until the index was
reached and the card was inserted. This rather bizarre behaviour was
needed in just one single place: refilling the piles in DealerScene::
setState. Reworking this code lets us get rid of this wonky API.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1107222 283d02a7-25f6-0310-bc7c-ecb5cbfe19da