Rewrite InteractorManager and Interactor.
The new version clearly separates the triggering logic from the
interaction logic. The new algorithms prioritize triggers that have been
active in related previous events, and allows to handle events with
multiple interactors at once (e.g., it is possible again to move the
viewport with the right button while one is dragging a piece with the
left button). Also, multiple triggers can now be assigned to one
interactor.

The rewrite temporarily sacrifices the possibility to load different
trigger setups from the config, but only to bring it back in a much
better fashion.

There is one known bug: Mouse interaction with the constraint visualizer
is not possible ATM, because the Interactors are filtering mouse events
quite aggressively lately.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1111390 283d02a7-25f6-0310-bc7c-ecb5cbfe19da