Remove onGameStateAlteredByUser() from KCardScene and DealerScene.
First off, this functionality is KPat specific, so it should have never
been in KCardScene in the first place. Secondly, it was really ugly. I
added this to KPat back when the KPat codebase was foreign and scary and
this seemed like the least dangerous way of adding the feature that I
needed (accurate tracking of whether the player had actually started
playing). Instead of special casing the two scenarios where card
movement shouldn't be considered starting the game, I decided to special
case every other card movement. It worked, but it was stupid and ugly.
Now it's gone. :)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1111482 283d02a7-25f6-0310-bc7c-ecb5cbfe19da