Major rewrite of pile layout code.
This is my second rewrite of the pile layout code. The first was just to
make it more readable. This one is to entirely change how it works,
especially handling more generic cases that we haven't actually run into
in KPat. Although it wasn't an explicit goal, I do think the new code is
more readable as well.

Previously a pile's reservedSpace and availableSpace where QSizeFs. Now
they're full blown QRectFs relative to the pile's position. This is much
more expressive and doesn't rely on interesting interpretations of
negative sizes.

Previously negatively aligned piles weren't really considered in the
contentSize calculations, meaning that their reservedSpaces often
overlapped those of other piles, which is obviously problematic. This is
no longer the case.

Previously, the first pile was grown as big as possible and then it
we checked to see who that conflicted with. But we always favoured the
first pile, giving it as much space a possible. Now we break the problem
down into four stages, growing the piles downward, upward, leftward and
rightward in separate steps. Also instead of giving the first pile as
much space as possible, we split the contested space between the two
piles. The result is a much fairer algorithm, although it certain
situations it still does favour earlier piles.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdegames@1118856 283d02a7-25f6-0310-bc7c-ecb5cbfe19da